split suffix file into more files
